+++ src/scripts/constants/item.ts
@@ -1,4 +1,5 @@
 export enum ItemTypes {
     WEAPON = 'weapon',
     ARMOUR = 'armour',
+    RING = 'ring',
 }