+++ src/scripts/constants/sprites.ts
@@ -87,5 +87,6 @@ export enum DungeonFeaturesSprites {
 export enum ItemSprites {
     WEAPON = 'weapon',
     ARMOUR = 'armour',
+    RING = 'ring',
 }
 export type DungeonTerrainSprites = DungeonFeaturesSprites & TerrainSprites;