+++ changelog.md
@@ -1,4 +1,5 @@
 ### 0.0.5 (currently developed)
+* (Feature)  Armours and weapons can be worn now, giving appropriate modifiers for stats and in combat
 * (Feature)  Added armours
 * (Refactor) Refactored game models so game data can be now easily serialized
 ### 0.0.4 (01.09.2019)